+++ src/Modules/Profile/ProfileGateway.php
@@ -6,15 +6,18 @@ use Foodsharing\Lib\WebSocketConnection;
 use Foodsharing\Modules\Core\BaseGateway;
 use Foodsharing\Modules\Core\Database;
 use Foodsharing\Modules\Core\DBConstants\Region\Type;
+use Foodsharing\Utility\WeightHelper;
 
 final class ProfileGateway extends BaseGateway
 {
 	private WebSocketConnection $webSocketConnection;
+	private $weightHelper;
 
-	public function __construct(Database $db, WebSocketConnection $webSocketConnection)
+	public function __construct(Database $db, WebSocketConnection $webSocketConnection, WeightHelper $weightHelper)
 	{
 		parent::__construct($db);
 		$this->webSocketConnection = $webSocketConnection;
+		$this->weightHelper = $weightHelper;
 	}
 
 	/**
@@ -358,6 +361,44 @@ final class ProfileGateway extends BaseGateway
 			':fs_id' => $fsId,
 			':date_from' => $this->db->date($from),
 			':date_to' => $this->db->date($to),
+		]);
+	}
+
+	public function getPickupsStat(int $fsId): array
+	{
+		$stm = 'SELECT 	DATE_FORMAT(a.date,\'%Y-%v\') AS calenderWeek,
+						bez.name AS districtName,
+						kat.name AS categorieName,
+						CASE b.abholmenge
+							WHEN 0 THEN \'' . $this->weightHelper->getFetchWeightName(0) . '\'
+							WHEN 1 THEN \'' . $this->weightHelper->getFetchWeightName(1) . '\'
+							WHEN 2 THEN \'' . $this->weightHelper->getFetchWeightName(2) . '\'
+							WHEN 3 THEN \'' . $this->weightHelper->getFetchWeightName(3) . '\'
+							WHEN 4 THEN \'' . $this->weightHelper->getFetchWeightName(4) . '\'
+							WHEN 5 THEN \'' . $this->weightHelper->getFetchWeightName(5) . '\'
+							WHEN 6 THEN \'' . $this->weightHelper->getFetchWeightName(6) . '\'
+							WHEN 7 THEN \'' . $this->weightHelper->getFetchWeightName(7) . '\'
+						END AS pickupAmount,
+						COUNT(*) AS pickupCount
+				FROM `fs_abholer` a
+					LEFT OUTER JOIN fs_betrieb b ON a.betrieb_id = b.id
+					LEFT OUTER JOIN fs_betrieb_kategorie kat ON b.betrieb_kategorie_id = kat.id
+					LEFT OUTER JOIN fs_bezirk bez ON b.bezirk_id = bez.id
+				WHERE a.foodsaver_id = :fs_id
+				  AND a.date > CURRENT_DATE() - INTERVAL 1 MONTH
+				GROUP BY DATE_FORMAT(date,\'%Y-%v\'),
+						 bez.name,
+						 kat.id,
+						 b.abholmenge
+				ORDER BY
+						a.`date` DESC,
+						bez.name ASC,
+						kat.id   ASC,
+						b.abholmenge ASC
+		';
+
+		return $this->db->fetchAll($stm, [
+			':fs_id' => $fsId,
 		]);
 	}